Refactor: support generating changelogs from commit messages between tags instead of milestones.
