+++ .gitignore
@@ -3,3 +3,5 @@ node_modules*
 npm-debug.log
 package-lock.json
 yarn.lock
+lib/
+lib.es
