+++ .npmignore
@@ -1,5 +1,3 @@
-typings
-tsd.json
 tsconfig.json
 .npmignore
 .gitignore
