+++ .travis.yml
@@ -1,8 +1,9 @@
 language: node_js
 node_js:
-  - 4
-  - 5
   - 6
+  - 7
+  - 8
+  - 9
 
 install:
   - npm install
