+++ index.ts
@@ -1,396 +1,38 @@
 #!/usr/bin/env node
 
-interface IConfig {
-  host: string
-  api: string
-  token: string
-}
-
-interface IMilestone {
-  id: number
-  iid: number
-  project_id: number
-  title: string
-  description: string
-  state: string
-  created_at: string
-  updated_at: string
-  due_date: any
-}
-
-interface IIssue {
-  id: number
-  iid: number
-  title: string
-  description: string
-  state: string
-  created_at: string
-  updated_at: string
-  labels?: string[]
-  milestone?: IMilestone
-  assignee?: any
-  author?: any
-}
-
-interface ILog {
-  version: string
-  content: string[]
-}
-
 /**
  * Module dependencies.
  */
-const fs = require('fs')
-const path = require('path')
-const _prompt = require('prompt')
-const jsonfile = require('jsonfile')
-const request = require('request')
-const semver = require('semver')
-const program = require('commander')
-const ProgressBar = require('progress')
+import { resolve } from 'path'
+import * as program from 'commander'
 
-/**
- * Default config.
- */
-const CONFIG = {
-  FILE: '.changelogrc',
-  HOST: 'gitlab.mogujie.org/',
-  API: 'http://gitlab.mogujie.org/api/v3',
-  OUTPUT: 'CHANGELOG.md'
-}
+import generateFromTags from './generateFromTags'
+import generateFromMilestones from './generateFromMilestones'
 
-const PROMPT = {
-  // description display at the top of console when prompting
-  DESC: `You have not configured it yet, have you?
-Please to work it out with the interactive prompt below.
-It will create a config file (.changelogrc) in your system.
+// absolute path to directory of lib
+const libPath = resolve(__dirname, '../')
 
-(If you have no idea about what token is, find it in your gitlab site by following "Profile Setting" - "Account" - "Reset Private token")
-
-Press ^C at any time to quit.
-`,
-
-  // options for prompt
-  OPTIONS: [
-    {
-      name: 'host',
-      message: 'Your gitlab host',
-      default: CONFIG.HOST
-    },
-    {
-      name: 'api',
-      message: 'Your gitlab api',
-      default: CONFIG.API
-    },
-    {
-      name: 'token',
-      message: 'Your private token'
-    }
-  ]
-}
-
-// absolute path of current project
+// absolute path of current gitlab project
 const root: string = process.cwd()
 
-// config file: ~/.changelogrc
-const configFileLocation: string = path.resolve(
-  process.env[process.platform === 'win32' ? 'USERPROFILE' : 'HOME'],
-  CONFIG.FILE
-)
-
 // CLI control
-program.version(require(`${root}/package`).version).parse(process.argv)
+program
+  .version(require(`${libPath}/package`).version, '-v, --version')
+  .option('-t, --tags', `Use tags instead of milestones`)
+  .parse(process.argv)
+
 main()
 
 // Main logic
 async function main() {
   try {
-    let config: IConfig
-
-    if (fsExistsSync(configFileLocation)) {
-      config = await getConfig()
+    if (program.tags) {
+      generateFromTags()
     } else {
-      config = await createConfigFile()
+      await generateFromMilestones()
     }
-
-    const { host, api, token } = config
-    const projectPath = getProjectPath(host)
-    const encodeProjectPath = encodeURIComponent(projectPath)
-    const tokenQuery = `private_token=${token}`
-
-    const milestonesApi = `${api}/projects/${encodeProjectPath}/milestones?${tokenQuery}`
-    const issuesApi = `${api}/projects/${encodeProjectPath}/milestones/#{milestoneId}/issues?${tokenQuery}`
-
-    const milestones: IMilestone[] = await fetchMilestones(milestonesApi)
-    const logs: ILog[] = await generateLogs(milestones, issuesApi)
-
-    // Generating changelog of current project
-    generateChangeLog(logs)
   } catch (e) {
     console.error(e)
+    process.exit(0)
   }
-}
-
-/**
- * Check weather file or directory exists on specific path
- * @param path
- */
-function fsExistsSync(path) {
-  try {
-    fs.accessSync(path, fs.F_OK)
-  } catch (e) {
-    return false
-  }
-  return true
-}
-
-function getConfig(): Promise<IConfig> {
-  return new Promise((resolve, reject) => {
-    jsonfile.readFile(configFileLocation, (e, data) => {
-      if (!e) {
-        resolve(data)
-      } else {
-        reject(e)
-      }
-    })
-  })
-}
-
-function createConfigFile(): Promise<IConfig> {
-  return new Promise((resolve, reject) => {
-    console.log(PROMPT.DESC)
-
-    _prompt.message = 'Enter'
-    _prompt.start()
-    _prompt.get(PROMPT.OPTIONS, (e, result) => {
-      if (e) {
-        reject(`\n${e}`)
-        return
-      }
-
-      const content: IConfig = {
-        host: result.host,
-        api: result.api,
-        token: result.token
-      }
-
-      jsonfile.writeFile(configFileLocation, content, e => {
-        if (!e) {
-          console.log('\n')
-          resolve(content)
-        } else {
-          reject(`Unable to write file: ${configFileLocation}`)
-        }
-      })
-    })
-  })
-}
-
-/**
- * Fetch all the milestones from the given api.
- *
- * @param  {string} api
- * @example of the api:
- * http://gitlab.mogujie.org/api/v3/projects/wxa%2Fwxa-module/milestones??per_page=10&private_token=xxx
- */
-function fetchMilestones(api: string): Promise<IMilestone[]> {
-  const promise = new Promise<IMilestone[]>((resolve, reject) => {
-    request(api, (e, response, body) => {
-      if (!e && response.statusCode === 200) {
-        const milestones: IMilestone[] = JSON.parse(body)
-        if (milestones.length) {
-          resolve(milestones)
-        } else {
-          reject(`There's no milestone yet.`)
-        }
-      } else if (response) {
-        reject(
-          `Unable to fetch milestones because: ${
-            JSON.parse(response.body).message
-          }.`
-        )
-      } else {
-        reject(e)
-      }
-    }).on('response', () => {
-      console.log('Starting to fetch all the milestones of this project.\n')
-    })
-  })
-
-  promise.then((response: Object[]) => {
-    if (response.length > 1) {
-      console.log(
-        `Get ${
-          response.length
-        } milestones totally.\nGetting start to fetch all the issues of these milestones:\n`
-      )
-    } else if (response.length === 1) {
-      console.log(
-        'Get only one milestone, fetch the issues of this milestone:\n'
-      )
-    }
-  })
-
-  return promise
-}
-
-/**
- * * Generating all the logs from milestones and the given api.
- *
- * @param  {IMilestone[]} milestones
- * @param  {string} api
- */
-async function generateLogs(
-  milestones: IMilestone[],
-  api: string
-): Promise<ILog[]> {
-  let promises: Promise<ILog>[] = milestones.map(milestone =>
-    generateLog(milestone, api)
-  )
-
-  const barOpts = {
-    complete: '=',
-    incomplete: ' ',
-    width: 20,
-    total: promises.length
-  }
-
-  const bar = new ProgressBar(
-    '  fetching issues [:bar] :percent :elapseds',
-    barOpts
-  )
-
-  promises.forEach(promise => {
-    promise.then(() => {
-      bar.tick()
-    })
-  })
-
-  return await Promise.all(promises)
-}
-
-/**
- * Generatin single log from specific milestone and the given api.
- *
- * @param  {IMilestone} milestone
- * @param  {string} api
- */
-function generateLog(milestone: IMilestone, api: string): Promise<ILog> {
-  return new Promise((resolve, reject) => {
-    request(
-      api.replace(/#{milestoneId}/, `${milestone.id}`),
-      (e, response, body) => {
-        if (!e && response.statusCode === 200) {
-          const version: string = milestone.title
-          const update: string = milestone.created_at.substr(0, 10)
-          const issues: IIssue[] = JSON.parse(body)
-
-          let content: string[] = issues.map(issue => {
-            if (issue.assignee && issue.assignee.username) {
-              return `- ${issue.title} (#${issue.iid} @${
-                issue.assignee.username
-              })`
-            } else {
-              return `- ${issue.title} (#${issue.iid})`
-            }
-          })
-
-          if (milestone.state !== 'closed') {
-            content.unshift(`## ${version}(unreleased)`)
-          } else {
-            content.unshift(`## ${version} - ${update}`)
-          }
-
-          resolve({
-            version: version,
-            content: content
-          })
-        } else if (response) {
-          reject(
-            `Unable to fetch issues of ${milestone.id} milestone because: ${
-              JSON.parse(response.body).message
-            }`
-          )
-        } else {
-          reject(e)
-        }
-      }
-    )
-  })
-}
-
-/**
- * Write the changelog into the output file.
- *
- * @param  {ILog[]} logs
- */
-function generateChangeLog(logs: ILog[]) {
-  console.log(`\nGenerating changelog into ${CONFIG.OUTPUT}`)
-
-  logs = logs.sort(compareVersions)
-  // console.log(logs)
-
-  let body: any = logs.map(log => {
-    return log.content.join('\n')
-  })
-
-  // console.log(body)
-  body = body.join('\n\n')
-  fs.writeFile(CONFIG.OUTPUT, body, e => {
-    if (!e) {
-      console.log(`\nOK, ${CONFIG.OUTPUT} generated successfully!`)
-    } else {
-      console.error(e)
-    }
-  })
-}
-
-/**
- * Get the full path of the current project.
- *
- * @param  {string} host
- * @return {string} projectPath
- */
-function getProjectPath(host: string): string {
-  let gitConfig: IConfig
-  let projectPath: string
-
-  try {
-    gitConfig = fs.readFileSync('.git/config', 'utf8')
-  } catch (e) {
-    throw `It can't be done because it's not a git project.`
-  }
-
-  try {
-    projectPath = `${gitConfig}`
-      .split(host)[1]
-      .split('\n')[0]
-      .replace(/(\:|\.git)/g, '')
-  } catch (e) {
-    throw `No gitlab project found in ${root}`
-  }
-
-  return projectPath
-}
-
-/**
- * Semver comparator
- *
- * @param  {ILog} log1
- * @param  {ILog} log2
- * @return {-1, 0, 1}
- *        Return 0 if v1 == v2,
- *        or 1 if v1 is greater,
- *        or -1 if v2 is greater.
- *        Sorts in ascending order if passed to Array.sort().
- */
-function compareVersions(log1: ILog, log2: ILog): number {
-  const v1: number = semver.clean(log1.version)
-  const v2: number = semver.clean(log2.version)
-
-  if (!v1 || !v2) {
-    return
-  }
-
-  return semver.rcompare(v1, v2)
 }
